First step of improving the quill editor
